More test cleaning
